rewrite lerna-scripts helpers in node
